backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# DB
sqlalchemy==2.0.30
alembic==1.13.1

# SQLite dev mode
aiosqlite==0.20.0

# Postgres prod mode (optional)
asyncpg==0.29.0
pgvector==0.2.5

# RAG / Embeddings
faiss-cpu==1.8.0
sentence-transformers==3.0.1
langchain==0.2.3
langchain-community==0.2.3

# LLM (Ollama)
ollama==0.2.1

# File parsing
pypdf==4.2.0
python-docx==1.1.0
markdown==3.6

# Observability
structlog==24.1.0
prometheus-client==0.20.0

# Tests
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# Misc
python-dotenv==1.0.1
httpx==0.27.0
numpy==1.26.4
